chore(create-component): rename `helpers` into `utils` (#1)
