+++ scripts/create-component/utils/appendToFile.js
