+++ scripts/create-component/utils/checkCwdRoot.js
