+++ scripts/create-component/utils/checkDirExists.js
