+++ scripts/create-component/utils/createFile.js
