+++ scripts/create-component/utils/getFilePath.js
