+++ scripts/create-component/utils/index.js
