+++ scripts/create-component/utils/mapFlagsToOptions.js
